renter,wallet,txpool: Update for core signature changes
